Move compile phase to StyleInjector for broad use
