+++ repos/re-theme/src/helpers/__tests__/compileStyles.js
@@ -19,6 +19,19 @@ const themeEvent = { fireThemeEvent: jest.fn(fireThemeEvent) }
 jest.setMock('../../theme/themeEvent', themeEvent)
 jest.setMock('../../dimensions', Dims)
 
+// tests helper that verifies dynamic keys aren't present in final styles object
+const assertNoDynamicKeys = styles => {
+  const styleKeys = Object.keys(styles)
+  expect(styleKeys).not.toContain('$xsmall')
+  expect(styleKeys).not.toContain('$small')
+  expect(styleKeys).not.toContain('$medium')
+  expect(styleKeys).not.toContain('$large')
+  expect(styleKeys).not.toContain('$web')
+  expect(styleKeys).not.toContain('$native')
+  expect(styleKeys).not.toContain('$ios')
+  expect(styleKeys).not.toContain('$android')
+}
+
 describe('Helpers', () => {
   describe('compileStyles', () => {
     afterEach(() => jest.clearAllMocks())
@@ -40,6 +53,8 @@ describe('Helpers', () => {
       expect(
         theme.button.contained.default.active.main.backgroundColor
       ).toEqual('orange')
+
+      assertNoDynamicKeys(theme)
     })
 
     it('should add the RTMeta key to the theme object when withMeta is true', () => {
@@ -90,19 +105,114 @@ describe('Helpers', () => {
       expect(compiled.components.button.default.padding).toEqual(
         testTheme.components.button.default.$web.padding
       )
+
+      assertNoDynamicKeys(compiled)
+    })
+
+    it('should compile size keys', () => {
+      const style = {
+        $xsmall: {
+          padding: 20,
+        },
+        $large: {
+          padding: 10,
+        },
+        $small: {
+          padding: 15,
+        },
+      }
+
+      const propsStyle = { margin: 12 }
+      const compiledStyles = compileStylesForState({ ...style, ...propsStyle}, sizeMap.hash.$large, 1000)
+      expect(compiledStyles).toEqual(
+        expect.objectContaining({
+          padding: style['$large'].padding,
+          ...propsStyle,
+        })
+      )
+      assertNoDynamicKeys(compiledStyles)
+    })
+
+    it('should compile platform keys', () => {
+      const style = {
+        $all: {
+          $large: {
+            borderTop: 10,
+            c: 'red',
+          },
+          $small: {
+            padding: 10,
+          },
+        },
+        $web: {
+          $xsmall: {
+            padding: 20,
+          },
+          $large: {
+            padding: 10,
+            c: 'blue',
+          },
+          $small: {
+            padding: 15,
+          },
+        },
+        $native: {
+          $large: {
+            padding: 0,
+          },
+        },
+      }
+      const propsStyle = { margin: 12 }
+      const compiledStyles = compileStylesForState({ ...style, ...propsStyle}, sizeMap.hash.$large, 1000)
+      expect(compiledStyles).toEqual(
+        expect.objectContaining({
+          padding: style.$web.$large.padding,
+          color: style.$web.$large.c,
+          borderTop: style.$all.$large.borderTop,
+          ...propsStyle,
+        })
+      )
+      assertNoDynamicKeys(compiledStyles)
     })
 
     it('should compile shortcuts', () => {
-      const compiled = compileStylesForState(
-        { p: 10, $small: { c: 'blue' }, $large: { c: 'green' } },
-        sizeMap.hash['$small'] + 10,
-        1000
+      const style = {
+        p: 20,
+        mT: 10,
+      }
+      const propsStyle = { mB: 12 }
+      const compiledStyles = compileStylesForState({ ...style, ...propsStyle}, sizeMap.hash.$large, 1000)
+      expect(compiledStyles).toEqual(
+        expect.objectContaining({
+          padding: style.p,
+          marginTop: style.mT,
+          marginBottom: propsStyle.mB,
+        })
       )
+      assertNoDynamicKeys(compiledStyles)
+      expect(Object.keys(compiledStyles)).not.toContain('mT')
+      expect(Object.keys(compiledStyles)).not.toContain('mB')
+      expect(Object.keys(compiledStyles)).not.toContain('p')
+    })
 
-      expect(compiled).toEqual({
-        padding: 10,
-        color: 'blue',
-      })
+    it('should maintain existing object structure', () => {
+      const style = {
+        component: {
+          p: 20,
+          mT: 10,
+        },
+      }
+      const propsStyle = { mB: 12 }
+      const compiledStyles = compileStylesForState({ ...style, ...propsStyle}, sizeMap.hash.$large, 1000)
+      expect(compiledStyles).toEqual(
+        expect.objectContaining({
+          component: {
+            padding: style.component.p,
+            marginTop: style.component.mT,
+          },
+          marginBottom: propsStyle.mB,
+        })
+      )
     })
   })
 })
