+++ repos/re-theme/src/reStyle/__tests__/reStyle.js
@@ -26,18 +26,7 @@ jest.setMock('../../styleInjector/styleInjector', {
 
 const { reStyle } = require('../reStyle')
 
-// tests helper that verifies dynamic keys aren't present in final styles object
-const assertNoDynamicKeys = styles => {
-  const styleKeys = Object.keys(styles)
-  expect(styleKeys).not.toContain('$xsmall')
-  expect(styleKeys).not.toContain('$small')
-  expect(styleKeys).not.toContain('$medium')
-  expect(styleKeys).not.toContain('$large')
-  expect(styleKeys).not.toContain('$web')
-  expect(styleKeys).not.toContain('$native')
-  expect(styleKeys).not.toContain('$ios')
-  expect(styleKeys).not.toContain('$android')
-}
+
 
 describe('reStyle', () => {
   afterEach(() => {
@@ -90,113 +79,4 @@ describe('reStyle', () => {
     expect(TestComp.props.style.margin).toBe(12)
   })
 
-  it('should compile size keys', () => {
-    const style = {
-      $xsmall: {
-        padding: 20,
-      },
-      $large: {
-        padding: 10,
-      },
-      $small: {
-        padding: 15,
-      },
-    }
-
-    const propsStyle = { margin: 12 }
-    const TestComp = reStyle(Component)(style).render({ style: propsStyle })
-    const compiledStyles = TestComp.props.style
-    expect(compiledStyles).toEqual(
-      expect.objectContaining({
-        padding: style['$large'].padding,
-        ...propsStyle,
-      })
-    )
-    assertNoDynamicKeys(compiledStyles)
-  })
-
-  it('should compile platform keys', () => {
-    const style = {
-      $all: {
-        $large: {
-          borderTop: 10,
-          c: 'red',
-        },
-        $small: {
-          padding: 10,
-        },
-      },
-      $web: {
-        $xsmall: {
-          padding: 20,
-        },
-        $large: {
-          padding: 10,
-          c: 'blue',
-        },
-        $small: {
-          padding: 15,
-        },
-      },
-      $native: {
-        $large: {
-          padding: 0,
-        },
-      },
-    }
-    const propsStyle = { margin: 12 }
-    const TestComp = reStyle(Component)(style).render({ style: propsStyle })
-    const compiledStyles = TestComp.props.style
-    expect(compiledStyles).toEqual(
-      expect.objectContaining({
-        padding: style.$web.$large.padding,
-        color: style.$web.$large.c,
-        borderTop: style.$all.$large.borderTop,
-        ...propsStyle,
-      })
-    )
-    assertNoDynamicKeys(compiledStyles)
-  })
-
-  it('should compile shortcuts', () => {
-    const style = {
-      p: 20,
-      mT: 10,
-    }
-    const propsStyle = { mB: 12 }
-    const TestComp = reStyle(Component)(style).render({ style: propsStyle })
-    const compiledStyles = TestComp.props.style
-    expect(compiledStyles).toEqual(
-      expect.objectContaining({
-        padding: style.p,
-        marginTop: style.mT,
-        marginBottom: propsStyle.mB,
-      })
-    )
-    assertNoDynamicKeys(compiledStyles)
-    expect(Object.keys(compiledStyles)).not.toContain('mT')
-    expect(Object.keys(compiledStyles)).not.toContain('mB')
-    expect(Object.keys(compiledStyles)).not.toContain('p')
-  })
-
-  it('should maintain existing object structure', () => {
-    const style = {
-      component: {
-        p: 20,
-        mT: 10,
-      },
-    }
-    const propsStyle = { mB: 12 }
-    const TestComp = reStyle(Component)(style).render({ style: propsStyle })
-    const compiledStyles = TestComp.props.style
-    expect(compiledStyles).toEqual(
-      expect.objectContaining({
-        component: {
-          padding: style.component.p,
-          marginTop: style.component.mT,
-        },
-        marginBottom: propsStyle.mB,
-      })
-    )
-  })
 })
